pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbx-sharepoint"
version = "0.1.0"
description = "Simplified SharePoint file and Excel interfaces for Azure Gov Databricks"
requires-python = ">=3.9"
dependencies = [
    "azure-identity>=1.12.0",
    "requests>=2.28.0",
    "pandas>=1.3.0",
    "openpyxl>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "responses>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dbx_sharepoint"]

[tool.pytest.ini_options]
testpaths = ["tests"]
